examples/sberbank-baseline/requirements.txt
ufal.udpipe==1.2.0.2
nltk==3.2.5
gensim==3.8.0
torch==1.1.0
catboost==0.16.5
pandas==0.24.2
pytorch_pretrained_bert==0.6.2
joblib==0.13.2
matplotlib==3.0.3
numpy==1.17.2
pymystem3==0.2.0
python-Levenshtein==0.12.0
sklearn_crfsuite==0.3.6
fastai==1.0.52
fuzzywuzzy==0.17.0
Keras==2.2.4
pymorphy2==0.8
scikit_learn==0.21.3
summa==1.2.0
